Added wraper to croper image
